feat: WeeklyReport

+++ app/src/main/java/com/example/composeapp/features/home/data/models/forecast/Current.kt
@@ -25,9 +25,9 @@ data class Current(
     @SerializedName("last_updated_epoch")
     val lastUpdatedEpoch: Int,
     @SerializedName("precip_in")
-    val precipIn: Int,
+    val precipIn: Double,
     @SerializedName("precip_mm")
-    val precipMm: Int,
+    val precipMm: Double,
     @SerializedName("pressure_in")
     val pressureIn: Double,
     @SerializedName("pressure_mb")
