+++ app/src/main/java/com/example/composeapp/features/home/data/models/forecast/Day.kt
@@ -11,7 +11,7 @@ data class Day(
     @SerializedName("avgtemp_f")
     val avgtempF: Double,
     @SerializedName("avgvis_km")
-    val avgvisKm: Int,
+    val avgvisKm: Double,
     @SerializedName("avgvis_miles")
     val avgvisMiles: Int,
     @SerializedName("condition")
@@ -37,11 +37,11 @@ data class Day(
     @SerializedName("mintemp_f")
     val mintempF: Double,
     @SerializedName("totalprecip_in")
-    val totalprecipIn: Int,
+    val totalprecipIn: Double,
     @SerializedName("totalprecip_mm")
-    val totalprecipMm: Int,
+    val totalprecipMm: Double,
     @SerializedName("totalsnow_cm")
-    val totalsnowCm: Int,
+    val totalsnowCm: Double,
     @SerializedName("uv")
     val uv: Int
 )
